Detect and skip bad H265 P-Frames (#6)
When an incoming stream is missing an I-Frame (full reference frame) and then a subsequent P-Frame (diff frame) comes in that references the missing one, the underlying hevc decoder will construct a blank gray reference frame to use instead, to gracefully recover.

In our use cases, we'd rather skip those, but they aren't exposed by the decoder when it happens, only via a log message.

So, to detect it, we install a log handler, detect that specific logger output, and then skip P-Frames until the next I-Frame (keyframe in AVFrame terms) comes in, then start returning frames again. It's not perfect but as long as no other part of the same program installs a custom log handler, it will work across an arbitrary number of FrameDecoder instances simultaneously within the same program without any extra overhead, since it uses a passed pointer to find it without lookup.

For other decoders, this has no effect. This is an imperfect solution but stops production of "unusable" (gray frame) data.

Signed-off-by: Emerson Knapp <emerson.b.knapp@gmail.com>